move templates and sampling rate to the previous position
