+++ utils/templates/templates.go
@@ -2,7 +2,6 @@ package templates
 
 import (
 	"github.com/netsampler/goflow2/v2/decoders/netflow"
-	"github.com/netsampler/goflow2/v2/state"
 )
 
 // Function that Create Template Systems.
@@ -11,5 +10,5 @@ type TemplateSystemGenerator func(key string) netflow.NetFlowTemplateSystem
 
 // Default template generator
 func DefaultTemplateGenerator(key string) netflow.NetFlowTemplateSystem {
-	return state.CreateTemplateSystem(key)
+	return netflow.CreateTemplateSystem(key)
 }
